+++ routes/breadcrumbs.php
@@ -3,11 +3,12 @@
 // Note: Laravel will automatically resolve `Breadcrumbs::` without
 // this import. This is nice for IDE syntax and refactoring.
 
-use App\Models\Category;
+use App\Models\Tag;
 use App\Models\Post;
 
 // This import is also not required, and you could replace `BreadcrumbTrail $trail`
 //  with `$trail`. This is nice for IDE type checking and completion.
+use App\Models\Category;
 use App\Services\Settings;
 use Diglactic\Breadcrumbs\Breadcrumbs;
 use Diglactic\Breadcrumbs\Generator as BreadcrumbTrail;
@@ -21,6 +22,12 @@ Breadcrumbs::for('category', function (BreadcrumbTrail $trail, Category $categor
     $trail->parent('home');
 
     $trail->push($category->title, route('category', $category));
+});
+
+Breadcrumbs::for('tag', function (BreadcrumbTrail $trail, Tag $tag) {
+    $trail->parent('home');
+
+    $trail->push($tag->title, route('tag', $tag));
 });
 
 Breadcrumbs::for('post', function (BreadcrumbTrail $trail, Category $category, Post $post) {